Refactored OneSignalStateSynchronizer
* Moved JSON diff generation logic into a new JSONUtil class.
* Move logic from OneSignalStateSynchronizer into UserStateSynchronizer to allow specific logic for push and email user records
* UserState inner class moved into it's own java file.
