+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/OSUtils.java
@@ -46,6 +46,7 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.security.MessageDigest;
 import java.util.Locale;
 import java.util.UUID;
 import java.util.regex.Pattern;
@@ -267,5 +268,20 @@ class OSUtils {
       } catch (JSONException e) {}
       
       return null;
+   }
+
+   static String hexDigest(String str, String digestInstance) throws Throwable {
+      MessageDigest digest = java.security.MessageDigest.getInstance(digestInstance);
+      digest.update(str.getBytes("UTF-8"));
+      byte messageDigest[] = digest.digest();
+
+      StringBuilder hexString = new StringBuilder();
+      for (byte aMessageDigest : messageDigest) {
+         String h = Integer.toHexString(0xFF & aMessageDigest);
+         while (h.length() < 2)
+            h = "0" + h;
+         hexString.append(h);
+      }
+      return hexString.toString();
    }
 }
