+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/SyncService.java
@@ -27,10 +27,7 @@
 
 package com.onesignal;
 
-import android.app.AlarmManager;
-import android.app.PendingIntent;
 import android.app.Service;
-import android.content.Context;
 import android.content.Intent;
 import android.os.IBinder;
 import android.support.annotation.Nullable;
